feat(generate): optimize top-k sampling
The transformers top-k sampling code is not optimized, as it applies the
softmax and multinomial sampling on the full [batch_size, vocab_size]
tensor, whereas it could apply it on the [batch_size, top-k] subset.
This cannot easily be optimized in transformers without breaking the
API, but we can provide an optimized neuron sampling loop.

Signed-off-by: David Corvoysier <david@huggingface.co>