feat(decoder): do not push checkpoint files to the hub
When serializing a model, the checkpoint files (i.e. the original
weights formatted for neuron) and the compiled artifacts are stored
under two folder in the model directory.
This modifies the NeuronForCausalLM.push_to_hub method to exclude
the folder containing the checkpoint files if they can be instead
reconstructed from a model on the hub.

Note that it is still possible to upload checkpoint files by using the
huggingface_cli directly on the saved model folder instead of
push_to_hub.

To allow optimum-neuron to fetch and reconstruct the checkpoint when
fetchin a neuron model from the hub, the original repository that was
used to export the model and its revision are stored in the neuron config.