Fix spelling in error message (#656)
I THINK that Talse should be False here.  The boolean flips depending on whether we are looking at inline_weights_to_neff or is_weights_neff_separated.  

It looks like the only time you would get to this is if inline_weights_to_neff is set to True, so the recommendation would b to switch it to False.