feat(llm): return all logits if batch size == 1
By default, transformers_neuronx model only return the logits for the
last input token, i.e. those leading to the generated token.
This modifies the default behaviour for models that support it (all but
gpt2), but only when batch_size is 1.
This does not have too much impact on performance and allows to evaluate
the model perplexity.