display available cached versions in error message
If a model is cached with a different configuration, I want to display alternative options to the user.  

If someone copies from the deploy code on Hugging Face and change something (e.g. sequence length), it is not obvious why it isn't working.

Based on a true story!

added some carriage returns to make it more readable
get_hub_cached_entries does generate an error if it is fed a model that doesn't have a model.  For example:  (randomly selected) model_id = "hexgrad/Kokoro-82M"

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/opt/aws_neuronx_venv_pytorch_2_1/lib/python3.10/site-packages/optimum/neuron/utils/hub_cache_utils.py", line 431, in get_hub_cached_entries
    model_type = target_entry.config["model_type"]
KeyError: 'model_type'

However, we call that function inside of is_cached, so I don't know if we are filtering for certain types of models or not.  If not, it should generate the error before it ever gets here