Fix training (#774)
# What does this PR do?

- [x] Updates the `NeuronTrainer` to match Transformers version
- [x] Fixes the way mixed-precision training is handled. It was breaking
multiple trainings such as Llama. We do not use `XLA_USE_BF16` and
`XLA_DOWNCAST_BF16` flags anymore following instructions from
[here](https://awsdocs-neuron.readthedocs-hosted.com/en/latest/general/appnotes/torch-neuronx/migration-from-xla-downcast-bf16.html)
- [x] Fixes support for gradient clipping: now it always happens between
the reduction of the gradients across the devices and the optimizer
step. The gradient norm is also now always reported when logging.
- [x] Adds support for `ignore_index` in the
`parallel_cross_entropy_loss`. There was a big issue in training when
using TP, the model was not learning. After investigating, it was linked
to the input being padded and the vanilla `parallel_cross_entropy` from
`neuronx_distributed` not supporting `ignore_index`:
- First, `loss.mean()` does not work in this case because the loss for
the ignored tokens is not zeroed.
- Second, the ignored tokens contributed to the gradient, which
effectively destroys training.


For now DP + TP can lead to compilation issues with SDK 2.20, but they
seem to be gone with SDK 2.21.

## Tests performed

- [x] Llama (`HuggingFaceTB/SmolLM2-135M-Instruct`) can overfit with
dp=1 tp=1
- [x] Llama (`HuggingFaceTB/SmolLM2-135M-Instruct`) + LoRA can overfit
with dp=1 tp=1
- [x] Llama (`meta-llama/Llama-3.2-1B`) can overfit with dp=1 tp=2
- [x] Llama + LoRA (`meta-llama/Llama-3.2-1B`) can overfit with dp=1
tp=2
- [x] Llama (`meta-llama/Llama-3.2-1B`) can overfit with dp=4 tp=2
**(Only tested on SDK 2.21, otherwise compiler error)**
- [x] Actual training of Llama (`meta-llama/Llama-3.2-1B`) dp=4 tp=2
**on SDK 2.21** and compared to GPUs

![W B Chart 20_02_2025
17_35_18](https://github.com/user-attachments/assets/ff1eefea-a82a-442a-bc8a-1deff3aa5ba6)

![W B Chart 20_02_2025
17_35_38](https://github.com/user-attachments/assets/8568351b-ef3c-49a0-893e-6a4aab9c05d6)

## To be done in following PRs

- Add `AdamW_FP32Params` for a more stable training in mixed-precision