revertme: disable continuous batching for phi3
For some reason, when using continuous batching, the prefill logits are
completely off, so we temporarily disable it.
The main consequence is that it is harder to fit a model compiled with a
high batch size, leading to either reducing it or using more cores.
For instance, with continuous batching on, we can use a batch size of 16
on 10 cores, while without we need to go down to a batch size of 8.