refactor(tests): remove tests that are skipped based on params
It makes sense to skip a test i the environment does not support it, but
it makes more sense to avoid having a test/parameters combination that
is then skipped. So here we remove tests parameters that are not
supported by test_parallelized_layers_model_matches_original.
This way it is easier to understand what tests are actually launched.