Initial PR for the documentation refactoring (#791)
# What does this PR do?

Related: #790 

This is the 1st PR for the refactoring of Optimum Neuron's
documentation, here below are the goals for this PR:

- [x] Re-structure the doc catalog
- [x] A representative model doc for each category
    - [x] transformers: text/vision/audio/multimodal; 
    - [x] diffusers; 
    - [x] sentence-transformer
* All docstring now moved to `optimum/neuron/utils/doc.py` -> according
changes in `modeling.py`, `modeling_diffusion.py`,
`modeling_seq2seq.py`.
* Modeling code for each representative under `optimum/neuron/models`,
these are more or less the same as previous `NeuronModelForXXX` but
crafted more for the specific model type according to their inputs.
Nothing new invented.
* Several bug fixes in the `pipeline` API while testing the code snippet
* CLIP
* Rename CLIP wrapper and neuron config (with projection layer it more
proper to name this way to distinguish from ClipVisionModel)
* Well separate using CLIP via transformers v.s. sentence transformers
- [ ] Inference modeling code for each model type, so far I just think
about replicating code as usually there is only `forward` to be
overridden for each class (if it becomes more complicated, we could
apply a similar mechanism to [Modular
Transformers](https://huggingface.co/docs/transformers/en/modular_transformers))

---------

Co-authored-by: Alvaro Moran <6949769+tengomucho@users.noreply.github.com>
Co-authored-by: David Corvoysier <david@huggingface.co>