Custom modeling for training (#801)
# What does this PR do?

### Custom modeling code for training

#### Features

This PR adds support for custom modeling code for training. 
Each custom modeling code can be added under
`optimum/neuron/models/training`.

Having a custom modeling allows us to implement Neuron specificities in
a cleaner way than using dynamic patching.
It becomes easy to:

- Fuse linear layers together for efficiency 
- Use custom linear layers such as `GQAQKVColumnParallelLinear`, useful
with high TP sizes.
- Use custom kernels, such as the flash attention kernel

In this PR we provide a first full custom implementation with Llama.

#### Model weight transformations

Because having a custom modeling code enables to change the vanilla
Transformers implementation, we need a way to make sure that we can load
checkpoints from Transformers, and that we can save checkpoints in the
original format as well.

To do that we provide an API with the `ModelWeightTransformationSpec`
classes.
These classes represent the transformation compared to the vanilla
Transformers implementation and are directly added in the modules
containing these transformations.

For now two exist:

- `FusedLinearsSpec`: represents a transformation when multiple linear
layers are fused into a single linear layer (possibly a parallel linear)
- `GQAQKVColumnParallelLinearSpec`: represents the transformation of
separate query, key, and value projections into a single
GQAQKVColumnParalleLinear projection.

Then during loading, saving and consolidation, we use these specs to
make sure every weight matches with Transformers weights.

#### Known issues

- There seems to be an issue when saving a checkpoint for DP > 1 during
training. After initial investigation, it seems to be a compiler bug,
but it will require more work. I suggest to work on it on a another PR.

### Training example

#### Specs

- Model: `meta-llama/Llama-3.2-3B-Instruct`
- Dataset: `databricks/databricks-dolly-15k`
- Trainer: `NeuronSFTTrainer`
- DP=4, TP=8
- Gradient accumulation steps = 16 => Effective batch size = 4 x 16 = 64
- Sequence length = 2048 with packing = True
- 3 epochs
- Learning rate = 5e-4, warmup ration = 0.3, lr scheduler type =
"cosine"

#### Loss curve

![W B Chart 24_04_2025
16_12_30](https://github.com/user-attachments/assets/cb4467cd-3e18-4560-86e0-d71d55bb08b7)


### To be done in later PRs:

- Support for PP
- Support for LoRA
- Refactor `save_pretrained` as it was done for `from_pretrained` in
this PR.
- Add test that tests overfitting