test(cache): modify traced entry condition
When running decoder cache tests, the generate_neff method is patched,
and it seems that just restoring it is not enough to avoid the
non-decoder tests to use also the patched version.
As a result, more local MODULES_xx files are present and the check for
the traced models must be slightly modified.