feat: Add Continuous pre-training example for SageMaker hyperpod (#842)
# What does this PR do?

<!--
Congratulations! You've made it this far! You're not quite done yet
though.

Once merged, your PR is going to appear in the release notes with the
title you set, so make sure it's a great title that fully reflects the
extent of your awesome contribution.

Then, please replace this with a description of the change and which
issue is fixed (if applicable). Please also include relevant motivation
and context. List any dependencies (if any) that are required for this
change.

Once you're done, someone will review your PR shortly (see the section
"Who can review?" below to tag some potential reviewers). They may
suggest changes to make the code even better. If no one reviewed your PR
after a week has passed, don't hesitate to post a new comment
@-mentioning the same persons---sometimes notifications get lost.
-->

<!-- Remove if not applicable -->

This PR adds a new example of using pre-built container to run
continuous pre-training using SageMaker Hyperpod. It is a new tutorial
under Tutorials -> Training Tutorials, similar to
https://huggingface.co/docs/optimum-neuron/en/training_tutorials/sft_lora_finetune_llm.

The content of setup the cluster refers to the Hyperpod workshop:
https://catalog.workshops.aws/sagemaker-hyperpod-eks/en-US/01-cluster.

Currently, the Dockerfile used in the example is build on top of the
previous SageMaker release image, once the new DLC is released, we will
replace it with the pre-built container.


## Before submitting
- [x] This PR fixes a typo or improves the docs (you can dismiss the
other checks if that's the case).
- [x] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?

---------

Co-authored-by: Pinak Panigrahi <panpinak@amazon.com>