refactor(decoder): use HF cache when loading checkpoints
Instead of downloading the checkpoint files explicitly when loading
a model checkpoint, we now try to take advantage of the Hugging Face
local cache. This is more efficient when you instantiate the same model
multiple times, for instance on an ec2 host.
This is also more efficient when using a shared model cache.