Remove example scripts (#893)
# What does this PR do?

There are a lot of example scripts taken from Transformers. First, they
do not work anymore, and I do not think that they serve a real purpose.
Removing them with the goal to replace them with less example scripts,
but well tested and relevant to what the library offers:

- Inference examples
- Training examples: regular models, distributed training, mutli-node
training. Basically these examples would be mapped to the tutorials in
the docs.