chore: remove doc-builder dependency from quality extra
To build the documentation, we use doc-builder, but having a git+https
dependency in the pyproject.toml file makes the pypi validation fail. so
far the solution was to manually remove the line before creating and
uploading the package, but this is not very clean nor practical.
So the dependency is removed, and the package is installed manually on
the pipelines that build the documentation.