refactor(nxd): remove neff_cache context
Patching the generate_neff method to instantiate and use a libneuronxla Cache
object to check for cached neffs is not required anymore, since the ModelBuilder
now does it also.
Note that we still need to use the hub_neuronx_cache context to patch the
libneuronxla create_compile_cache method and use the hf hub cache proxy instead
of the local cache.