Trainer simplification (#938)
# What does this PR do?

This PR takes the `Trainer` from tranformers and tries to maintains most
of its most important features and adapt them to work on AWS Trainium
via the `NeuronTrainer`. Prior to this PR we would inherint from the
`Trainer`, now we are completely disconnected from it, having full
control on what we change and when we change it, as well as removing a
lot of code that is simply not relevant to us for Neuron.

We also define our own independent `NeuronTrainingArguments` in the same
fashion.

For TRL Trainers, we simply adapt the `NeuronSFTTrainer` to work with
it, we might apply the same philosophy in a later PR where we will not
rely on the TRL library, let's see if it makes sense. We remove the
`NeuronORPOTrainer` that was implemented only for the non-distributed
setting and was not maintained.

Overall this PR simplifies a lot of the code and should make maintenance
easier in the future.

Left to be done / checked in later PR:

- [ ] Test and check that ZeRO-1 works for DP, DP + TP, PP and with PEFT
- [ ] Test and check that gradient checkpointing works properly
- [ ] Solve the cast to bf16 approach instead of providing the
`torch_dtype`?
- [ ] Add metrics to the trainer: # token seen in training, #tokens per
seconds, mfu, etc
- [ ] Add resume training feature
- [ ] Update to latest TRL version and update the `NeuronSFTTrainer`
accordingly.
- [ ] Add master weight saved in optimizer / optimizer state in FP32