ZeRO-1 and mixed-precision (#956)
# What does this PR do?

It adds a better support for mixed-precision training. The following
cases are supported:

- No ZeRO-1 optimizer, full bf16 model weights. By default in the
`NeuronTrainer` there is no support for fp32 grads or optimizer states
in this case. It can be done manually.
- No ZeRO-1 optimizer, `torch.autocast`. In this case the optimizer
states will be in fp32, and only the relevant ops will be casted
automatically for the forward / backward by `torch.autocast`.
- ZeRO-1 optimizer. In this case we support: bf16 / autocast weights
with fp32 grads and optimizer states

On top of this, we refactor the tests at the same time Instead of having
a `test_common.py` file, we now have:

- `test_checkpointing.py` for checkpointing tests
- `test_zero_1.py` for ZeRO-1 tests
- `test_optimizer.py` for optimizer tests that do not use ZeRO-1
optimization
- `test_mixed_precision.py` for mixed precision tests