ci(venv): configure pytorch installation
Note that github variables used for inputs can only be of type string.
This is why the 'use_cuda' variable is not a boolean.
Being able to configure the pytorch installation allows a specific
workflow to install a specific torch version, or to use CUDA (some
packages are not compatible with pytorch CPU version).