fix(vllm): update ModelConfig patching
With vLLM 0.11.0, the hard-coded ModelConfig verification with parallel config
happens twice:
- when the ModelConfig is created (and after the first neuron platform
  callback has been called),
- when the ModelConfig is used, in a different process than the one used
to create that config (and after the second callback thas been called to
update the vLLM config).
The existing dynamic patch in the first callback only inhibits the first
check: another dynamic patch is added in the second platform callback,
patching the ModelConfig instance instead of patching the ModelConfig class.