feat(vllm): serve command offers dash/underscore CLI option aliases
vLLM has a rather complex argument parsing system that ends up
converting options from dashes to underscore. This allows to use, for
example, --served-model-name and --served_model_name without noticing.
The "serve" subcommand in Optimum Neuron is parsed with ArgumentParser
without these fancy transformation, but we can still provide an alias to
allow options to be passed with dashes, making them interchangeable with
the options that would be passed to vLLM.